add base64 and yaw, pitch

+++ testPCLDll/PclRenderer.cpp
@@ -38,9 +38,10 @@ void Render(PclRendererInterface* _obj){
 
 void RenderOffScreen(PclRendererInterface* _obj){
     _obj->RenderOffScreen();
+    
 }
 
-void MoveCamera(PclRendererInterface* _obj,int _x,int _y){
+void MoveCamera(PclRendererInterface* _obj,double _x,double _y){
     _obj->MoveCamera(_x,_y);
 }
